OpenAI_SDK_Agents/12_custom_runner/pyproject.toml
[project]
name = "12-runner-class"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai-agents>=0.2.8",
]
